requirements.txt
PyGithub
python-gitlab
rdflib